Add CEDDL out-of-the-box ruleset; test against this ruleset
Previously, CEDDL tests were against an example ruleset that
didn't match the out-of-the-box ruleset provided for organizations
which enabled the DLO integration and toggled on the Customer
Experience Digital Data Layer: 1.0 option. Now, we test against the
default ruleset. This required removing and updating a number of
tests since some expectations were no longer valid.